Add test to ensure auth is called before uninstall (#124)

+++ vscode/src/toitUninstall.ts
@@ -3,10 +3,11 @@
 // found in the LICENSE file.
 
 import { App, RelatedApp } from "./app";
-import { Context, uninstallApp } from "./utils";
+import { Context, ensureAuth, uninstallApp } from "./utils";
 
 export function createUninstallCommand(ctx: Context): (app: App) => void {
   return async(app: RelatedApp) => {
+    if (!await ensureAuth(ctx)) return;
     await uninstallApp(ctx, app.app());
     ctx.refreshDeviceView();
   };
